@@ -1,8 +1,9 @@
 {
   "name": "cytoscape-elk",
-  "version": "1.1.10",
+  "version": "2.0.0",
   "description": "The ELK layout algorithm adapter for Cytoscape.js",
-  "main": "cytoscape-elk.js",
+  "main": "dist/cytoscape-elk.js",
+  "module": "src/index.js",
   "author": {
     "name": "John Stephenson"
   },
@@ -10,18 +11,16 @@
     "gh-pages:demo": "cpy demo.html . --rename=index.html",
     "gh-pages:deploy": "gh-pages -d . -v node_modules/**",
     "gh-pages:clean": "rimraf index.html",
-    "copyright": "update license",
     "lint": "eslint src",
-    "build": "cross-env NODE_ENV=production webpack",
-    "build:min": "cross-env NODE_ENV=production MIN=true webpack",
-    "build:release": "run-s build copyright",
-    "watch": "webpack --progress --watch",
+    "build": "webpack --mode=production",
+    "build:release": "run-s build",
+    "watch": "webpack --progress --watch --mode=development",
     "dev": "webpack-dev-server --open",
     "test": "mocha"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/jfstephe/cytoscape.js-elk.git"
+    "url": "https://github.com/cytoscape/cytoscape.js-elk.git"
   },
   "keywords": [
     "cytoscape",
@@ -29,32 +28,32 @@
   ],
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/jfstephe/cytoscape.js-elk/issues"
+    "url": "https://github.com/cytoscape/cytoscape.js-elk/issues"
   },
-  "homepage": "https://github.com/jfstephe/cytoscape.js-elk",
+  "homepage": "https://github.com/cytoscape/cytoscape.js-elk",
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-preset-env": "^1.5.1",
-    "camelcase": "^4.1.0",
-    "chai": "4.0.2",
-    "cpy-cli": "^1.0.1",
-    "cross-env": "^5.0.0",
-    "eslint": "^3.9.1",
-    "gh-pages": "^1.0.0",
-    "mocha": "3.4.2",
-    "npm-run-all": "^4.1.2",
-    "rimraf": "^2.6.2",
-    "update": "^0.7.4",
-    "updater-license": "^1.0.0",
-    "webpack": "^2.6.1",
-    "webpack-dev-server": "^2.4.5"
+    "@babel/core": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "babel-loader": "^8.0.0",
+    "camelcase": "^5.3.1",
+    "chai": "4.2.0",
+    "cpy-cli": "^3.0.0",
+    "eslint": "^6.8.0",
+    "gh-pages": "^2.2.0",
+    "mocha": "7.0.0",
+    "prettier": "^1.19.1",
+    "rimraf": "^3.0.0",
+    "webpack": "^4.41.5",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.10.1"
   },
   "peerDependencies": {
     "cytoscape": "^3.2.0"
   },
   "dependencies": {
-    "clean": "^4.0.2",
-    "elkjs": "^0.3.0"
+    "elkjs": "^0.5.1"
+  },
+  "prettier": {
+    "singleQuote": true
   }
 }
